requirements.txt
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
llama-index-vector-stores-elasticsearch
llama-index-vector-stores-chroma
llama-index-llms-ollama
llama-index
openai
llama-index
datasets
transformers
cohere
pypdf
Markdown
docx2txt
llama-index-readers-file

pydantic==1.10.11
llama-index-embeddings-huggingface
llama-index-embeddings-instructor
sentence-transformers
chromadb~=0.4.24
python-dotenv~=1.0.1